locator: remove unused cursor struct
this struct was initially provided to enable to the locator to
serialize/deserialize control plane cursors, and determine if
there are subsequent pages to fetch.

however, now that there is a `has_more` field in control plane payload
this is no longer needed. this is a better solution as synapse
is now agnostic to getsentry cursor formats, and should not needed to
understand them at all. so this code is no longer needed.