locator: remove unused cursor struct (#32)
this struct was initially provided to enable the locator to serialize/deserialize control plane cursors, and determine if there are subsequent pages to fetch.

however, now that there is a `has_more` field in the control plane payload, so this logic is no longer needed. this is a better solution as synapse is now agnostic to getsentry cursor formats, and should not need to understand their formats at all. so this code can be deleted.