locator: deprecate testing route provider
now that the filesystem based route provider has been implemented, we can
use that in tests and deprecate the TestingRouteProvider. the testing
route provider is less realistic and we have no more reason to keep
that code around even for tests.

also updates the example configs so the filesystem version is used in local dev