locator: add high-level structure for incremental lookup and negative caching
- implement worker loop to periodically fetch results and handle any refresh / shutdown
command received from other threads. refresh interval is currently hardcoded to between 1 and 60
seconds, but these will need to be tested and potentially modified or made configurable
later depending on our findings.

- introduce negative cache module to temporarily store missed keys. it's purpose is to
avoid repeated lookups for the same (non-existing) organization key, and thus reduce load
on code/systems behind this cache. caching is very lightweight with max size of 1000 records
and ttl of 5 seconds, and is only designed to avoid very excessive load.

- make locator.lookup() and associated functions async. since they potentially have to wait
for result from the other thread via the channel receiver if there is a cache miss

- remove `refresh()` from the public locator API -- this shouldn't be called by end users

- incremental_load() implementation is still pending and will be added in a next step