feat(ingest protocol): Add code to work with relay protocol
Added a submodule to handle everything related to project configs.
This is done because the code needed to handle proeject configs is
pretty large and I would like to isolate that part of the code.

The first thing added in the sub module is the protocol used to
handle the project configs. It has the request and response payload.
Fields which would be modified by the router are defined in the
appropriate structs. Everything else would go in extra_fields which
would be a catch all. This would allow forward compatibility with
future field additions and removals.