ingest-router: add cell validation on project configs (#91)
this change ensures that locality boundaries are enforced by synapse on
the project configs endpoint
e.g. a us org cannot be requested via the de domain and the s4s2 org
cannot be requested via the us or de domains. this guarantees proper
isolation of each locality and single-tenant even if they share the
edge/control silo.