ingest-router: Simplify handler type
This chnage simplifies the handler trait so it is no longer generic.

The generics solution does not work well for our use case since
generics requires the concrete type to be known at compile time,
however in our case the router picks the handler dynamically at
runtime based on inspecting the incoming route.

With Arc<dyn Handler>, we can more easily dyanamically dispatch to
the appropriate handler based on the matched route.

Each handler receives the raw requests/responses and is responsible
for splitting and merging them as needed.