ingest-router: Simplify handler type (#92)
This change does a few things:
- simplifies the handler trait so it is no longer generic, and enables
additional handlers to be added
- adds a "health" handler stub -- this doesn't do anything yet except
demonstrates that it is now possible
to now add more than one handler.
- finishes the merge responses implementation in the project configs
handler

The generics solution does not work well for our use case since it
requires the concrete type to be known at compile time. In our case the
router picks the handler dynamically at runtime based on inspecting the
incoming route.

With `Arc<dyn Handler>`, we can more easily dyanamically dispatch to the
appropriate handler based on the matched route.

Each handler receives the raw requests/responses and is responsible for
splitting and merging them as needed.